component name changed and adding callback
